DermaFast/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
bcrypt==4.1.2
httpx>=0.24.0,<0.25.0